meiduo_mall/requirements.txt
aliyun-python-sdk-core-v3==2.8.9
aliyun-python-sdk-dysmsapi==1.0.0
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
Django==1.11.11
django-cors-headers==2.3.0
django-filter==2.0.0
django-redis==4.9.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
npm==0.1.1
optional-django==0.1.0
Pillow==5.2.0
pkg-resources==0.0.0
pycparser==2.18
pycryptodome==3.6.4
PyJWT==1.6.4
PyMySQL==0.9.2
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
